feat(sbt-plugin): announce support for proto3 optional fields
It doesn't seem like we need to treat them specially, we can just 'say' we support them.

I'm not sure we need to do anything at the Maven side, will check that next.
